% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FI.GDSC.ANOVA.Statistics_library.R
\name{gdscANOVA_cohens_d}
\alias{gdscANOVA_cohens_d}
\title{Effect Size through Cohen's D}
\usage{
to edit
}
\description{
\code{gdscANOVA_cohens_d(x,y)} returns the effect size of the interaction between the values in the set given by the union of x and y and the factor inducing the dychotomy \eqn{{x,y}} onto it
}

