% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FI.GDSC.ANOVA.Statistics_library.R
\name{gdscANOVA_individualANOVA}
\alias{gdscANOVA_individualANOVA}
\title{Individual ANOVA test}
\usage{
gdscANOVA_individualANOVA(DRUG_ID, FEATURE, display=TRUE, printTOfig=FALSE, PATH='', FN='', FDR=NA, OUTPUT_PATH='')
}
\arguments{
\item{DRUG_ID}{A string containing the identifier of the drug, whose IC50s must be tested (i.e. must be a row name of the \code{IC50s} matrix)}

\item{FEATURE}{}

\item{display}{}

\item{printTOfig}{}

\item{PATH}{}

\item{FN}{}

\item{FDR}{}

\item{OUTPUT_PATH}{}
}
\value{
A
}
\description{
\code{gdscANOVA_individualANOVA(DRUG_ID, FEATURE, display=TRUE, printTOfig=FALSE, PATH='', FN='', FDR=NA, OUTPUT_PATH='')} \cr\cr
executes an individual ANOVA test to evaluate the statistical interaction between the IC50 values of a given drug, across cell lines, and an underlying
binary factor. \cr\cr
Calling this function requires the existance of the following global variables, described in \url{https://github.com/francescojm/FI.GDSC.ANOVA/wiki} and created by the functions in
\code{FI.GDSC.ANOVA.Preamble_Library.R}: \cr
\itemize{
  \item IC50s = Matrix with IC50 values
  \item InputFeatures = List containing the input features to be correlated with drug response
  \item GDSCANOVA_SETTINGS = List of settings variables
}
}
\author{
Francesco Iorio - \email{iorio@ebi.ac.uk}
}
\seealso{
A
}

