% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FI.GDSC.ANOVA.Preamble_Library.R
\name{gdscANOVA_Preamble}
\alias{gdscANOVA_Preamble}
\title{ANOVA preamble}
\usage{

  gdscANOVA_Preamble(ANOVA_setting_file = "HTML_templates_and_elements/FI.GDSC.ANOVA.Settings_file_template.csv")
}
\description{
\code{gdscANOVA_Preamble(ANOVA_setting_file)} returns a list of setting variables
initilized with the values contained in the ANOVA_settig_file [see an example of this file
in the HTML_templates_and_elements/FI.GDSC.ANOVA.Settings_file_template.csv]
}

