% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FI.GDSC.ANOVA.Preamble_Library.R
\name{gdscANOVA_Preamble_loading}
\alias{gdscANOVA_Preamble_loading}
\title{ANOVA preamble loading}
\usage{
gdscANOVA_Preamble_loading(global_setting_list)
}
\description{
\code{gdscANOVA_Preamble_loading(global_setting_list)} loads and export in the workspace, as global variables
objects and list, whose path is contained in the global_setting_list
}

