% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FI.GDSC.ANOVA.Statistics_library.R
\name{gdscANOVA_individualANOVA}
\alias{gdscANOVA_individualANOVA}
\title{Individual ANOVA test
Calling this function requires the existance of the following global variables
(created through the functions in Data_Manipulation_Library.R, Preamble.R, or loaded from relevant R objects):
           IC50s = Matrix with IC50 values
           InputFeatures = Structure containing the input features to be correlated with drug response
           gdscANOVA.settings.featFactorPopulationTh = How many sample that are positive for a given feature must be present in order for the test to be performed
           gdscANOVA.settings.includeMSI_Factor = Boolean variable specifying if the Microsatellite instability status of the samples should be included as co-factor
           gdscANOVA.settings.MSIfactorPopulationTh = How many MicroSatellite instable samples must be present in order for the test to be performed}
\usage{
gdscANOVA_individualANOVA(DRUG_ID, FEATURE, display = TRUE,
  printTOfig = FALSE, PATH = "", FN = "", FDR = NA, OUTPUT_PATH = "")
}
\description{
Individual ANOVA test
Calling this function requires the existance of the following global variables
(created through the functions in Data_Manipulation_Library.R, Preamble.R, or loaded from relevant R objects):
           IC50s = Matrix with IC50 values
           InputFeatures = Structure containing the input features to be correlated with drug response
           gdscANOVA.settings.featFactorPopulationTh = How many sample that are positive for a given feature must be present in order for the test to be performed
           gdscANOVA.settings.includeMSI_Factor = Boolean variable specifying if the Microsatellite instability status of the samples should be included as co-factor
           gdscANOVA.settings.MSIfactorPopulationTh = How many MicroSatellite instable samples must be present in order for the test to be performed
}

